pyproject.toml
[project]
name = "temporal-video-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.128.1",
    "google-generativeai>=0.8.6",
    "motor>=3.7.1",
    "opencv-python>=4.12.0.88",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.15.5",
    "python-jose[cryptography]>=3.5.0",
    "scikit-learn>=1.8.0",
    "slowapi>=0.1.9",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.248",
    "uvicorn>=0.40.0",
]
